refactor code and remove containers
